requirements.txt
fastapi
uvicorn
python-dotenv
email-validator
jinja2
